cert_api/requirements.txt
flask
prometheus_client
psycopg2-binary